Custom PCB Shield
This custom PCB is a shield for the Arudino Mega.  It takes 24V from the power supply (I included 2 holes below the solder pads to zip tie the wires in place) and uses 2 DC-DC step-down converters (https://www.amazon.com/gp/product/B01MQGMOKI/ref=oh_aui_detailpage_o03_s00?ie=UTF8&psc=1), one for 12V to power the Arduino Mega, and 5V to power the Raspberry Pi.  If you are not using the Pi, do not worry about the 5V converter.  I included space for 3-pin headers immediately next to each step-down converter for voltage monitoring - this can be for setting the voltage initially or can be implemented more permanently.  I'm using these displays, but any similar display should work: https://www.amazon.com/gp/product/B077ZQQQFC/ref=oh_aui_detailpage_o02_s00?ie=UTF8&psc=1.  Another solution is to just use a standard multimeter to set the output voltage.

The RPi can be directly soldered at the top of the board, or a 2x20 female pin header (similar to the 2x18 header on the Mega) can be soldered for a less permanent installation.  I got my RPi Zero W at Microcenter for $5, so I may just make it permanent.

The remainder of the I/O is with screw terminals.  The terminals are a 0.1" / 2.54mm pitch.  I got a combination of 5-, 7-, and 8-connector headers since larger ones get pricey.  I purchased them on Arrow.com for pretty cheap.

Finally, the most difficult part of the board is soldering the resistors.  I used SMD resistors (0805 package size) and I'm planning to solder them with a fine tip on my soldering iron.  Alternatively, they can be soldered using solder paste and by putting the board in a toaster oven (one which you're not planning to cook food in ever again!).  If enough people would prefer, I can use the blank real estate on the left of the board to incorporate through-hole resistors instead.  To be clear, this is exclusively so I don't need to run 3 wires to the limit switches - I'm only running 5V and signal to each limit switch, and using the signal tied to ground with a pull-down resistor on the PCB.

For those interested in using the RPi, I will continue to validate the code updates on my end and will push them out here once I know it all works.